+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/export/_base.py
@@ -58,6 +58,9 @@ class BaseExporter:
         """Azure Monitor base exporter for OpenTelemetry.
 
         :keyword str api_version: The service API version used. Defaults to latest.
+        :keyword str connection_string: The connection string used for your Application Insights resource.
+        :keyword bool enable_local_storage: Determines whether to store failed telemetry records for retry. Defaults to `True`.
+        :keyword str storage_path: Storage path in which to store retry files. Defaults to `<tempfile.gettempdir()>/opentelemetry-python-<your-instrumentation-key>`.
         :rtype: None
         """
         parsed_connection_string = ConnectionStringParser(kwargs.get('connection_string'))
@@ -107,6 +110,8 @@ class BaseExporter:
                 name="{} Storage".format(self.__class__.__name__),
                 lease_period=self._storage_min_retry_interval,
             )
+        # specifies whether current exporter is used for collection of instrumentation metrics
+        self._instrumentation_collection = kwargs.get('instrumentation_collection', False)
         # statsbeat initialization
         if self._should_collect_stats():
             # Import here to avoid circular dependencies
@@ -285,7 +290,8 @@ class BaseExporter:
     def _should_collect_stats(self):
         return is_statsbeat_enabled() and \
             not get_statsbeat_shutdown() and \
-            not self._is_stats_exporter()
+            not self._is_stats_exporter() and \
+            not self._instrumentation_collection
 
     # check to see if statsbeat is in "attempting to be initialized" state
     def _is_statsbeat_initializing_state(self):