+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Add validation logic to ApplicationInsightsSampler
     ([#26546](https://github.com/Azure/azure-sdk-for-python/pull/26546))
+- Add pre-aggregated standard metrics - requests/duration, dependencies/duration
+    ([#26546](https://github.com/Azure/azure-sdk-for-python/pull/26546))
 
 ### Breaking Changes